[Solana]: Move instruction builders to dedicated modules
